requirements.txt
pandas
numpy
ccxt
scikit-learn
xgboost
python-dotenv
ta
joblib